+++ src/tracee/util.h
@@ -22,7 +22,7 @@ struct PidTid
 };
 
 u64 *register_by_number(user_regs_struct *regs, int reg_number) noexcept;
-u64 get_register(user_regs_struct *regs, int reg_number) noexcept;
+u64 get_register(user_regs_struct *regs, size_t reg_number) noexcept;
 u64 GetDwarfRegister(const u64 *registerCache, size_t number) noexcept;
 size_t GetDwarfRegisterIndex(size_t dwarfNumber) noexcept;
